# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tes E-commerce Simple API",
        "description": "Dokumentasi API untuk tes e-commerce sederhana dengan Laravel dan L5-Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9090",
            "description": "Localhost API server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register user baru",
                "operationId": "92af7d5f95c73e20136cf817e47bde35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "budi@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil register"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "9c7361a02ab87c6f7609f2293c9abf69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "budi@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil login"
                    },
                    "401": {
                        "description": "Login gagal (email/password salah)"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Error server"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "6c9562d211d58e878cfd7d1c68aa02c8",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    },
                    "401": {
                        "description": "Tidak terautentikasi / token tidak valid"
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token hasil login: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}